Add more detailed errors and tests in the windows/service metricset (#17725)
* error details, tests

* changelog

* work on tests

* refactoring

* mage fmt

* mage fmt

(cherry picked from commit c51a3bf29e65eda018bab7836d0bdd15f054932b)
