+++ metricbeat/module/windows/service/service.go
@@ -37,14 +37,14 @@ func init() {
 // multiple fetch calls.
 type MetricSet struct {
 	mb.BaseMetricSet
-	reader *ServiceReader
+	reader *Reader
 }
 
 // New create a new instance of the MetricSet
 // Part of new is also setting up the configuration by processing additional
 // configuration entries if needed.
 func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
-	reader, err := NewServiceReader()
+	reader, err := NewReader()
 	if err != nil {
 		return nil, err
 	}
